pyproject.toml
[tool.poetry]
name = "learnmate-ai-powered-study-partner"
version = "0.1.0"
description = "LearnMate is an AI-driven study assistant that personalizes learning experiences for students."
authors = ["Anurag Singh <singhanurag024@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.95.0"
uvicorn = "^0.21.0"
pydantic = "^1.10.7"
python-dotenv = "^1.0.1"
openai = "^0.27.0"
aiohttp = "^3.10.5"
elevenlabs = "^1.7.0"
sqlmodel = "^0.0.22"

[tool.poetry.dev-dependencies]
pytest = "^7.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"